limited update to corresponding MATLAB classes and tests

+++ test/unit_tests/TestNormalBayesClassifier.m
@@ -6,7 +6,7 @@ classdef TestNormalBayesClassifier
     methods (Static)
         function test_1
             X = [randn(50,3)+1;randn(50,3)-1];
-            Y = [ones(50,1);-ones(50,1)];
+            Y = int32([ones(50,1);-ones(50,1)]);
             classifier = cv.NormalBayesClassifier;
             classifier.train(X,Y);
             Yhat = classifier.predict(X);
